# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Daliys Analytics Server — SDK & Dashboard API
  version: "1"
  description: >
    Authoritative wire contract for the analytics server. The SDK-facing
    /v1 endpoints below are fully specified per
    server_implementation_plan.md section 5 — any change here must update
    contracts/fixtures, internal/contracts, and the Unity SDK together.
    The /api/v1 dashboard endpoints are specified through Overview/Event
    Explorer/auth (Phase S2); funnel/retention/timeline/catalog/system
    remain path stubs until Phase S3.
servers:
  - url: https://analytics.example.com

components:
  securitySchemes:
    installationBearer:
      type: http
      scheme: bearer
      description: The installation_credential returned implicitly by registration.
    dashboardSession:
      type: apiKey
      in: cookie
      name: session

  schemas:
    ClientPolicy:
      type: object
      required: [mode, next_check_seconds, discard_pending]
      properties:
        mode:
          type: string
          enum: [active, pause_upload, disable_collection]
        next_check_seconds:
          type: integer
          minimum: 1
        discard_pending:
          type: boolean

    ErrorResponse:
      type: object
      required: [server_time, code, message, request_id]
      properties:
        server_time: { type: string, format: date-time }
        code: { type: string }
        message: { type: string }
        request_id: { type: string }

    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - project_id
        - install_id
        - installation_credential
        - sdk_name
        - sdk_version
        - app_version
        - build_number
        - platform
      properties:
        schema_version: { type: integer, const: 1 }
        project_id: { type: string }
        install_id: { type: string, format: uuid }
        installation_credential:
          type: string
          description: 32 cryptographically random bytes, unpadded base64url.
        sdk_name: { type: string }
        sdk_version: { type: string }
        app_version: { type: string }
        build_number: { type: string }
        platform: { type: string }

    RegisterResponse:
      type: object
      required: [server_time, installation_status, client_policy]
      properties:
        server_time: { type: string, format: date-time }
        installation_status: { type: string, enum: [registered] }
        client_policy: { $ref: "#/components/schemas/ClientPolicy" }

    EventProperties:
      type: object
      description: >
        At most 32 flat keys. Values are string (<=1024 UTF-8 bytes),
        finite number, boolean, or null. Arrays and nested objects are
        rejected. Encoded properties are at most 8 KiB total.
      additionalProperties:
        oneOf:
          - type: string
          - type: number
          - type: boolean
          - type: "null"

    Event:
      type: object
      additionalProperties: false
      required:
        - event_id
        - session_id
        - sequence
        - session_elapsed_ms
        - name
        - occurred_at_client
        - app_version
        - build_number
        - platform
        - os_version
        - device_class
        - locale
        - timezone_offset_minutes
        - properties
      properties:
        event_id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        sequence: { type: integer, minimum: 0 }
        session_elapsed_ms: { type: integer, minimum: 0 }
        name:
          type: string
          pattern: "^[a-z][a-z0-9_]{0,63}$"
          description: >
            sys_ prefix is reserved for SDK-owned system events
            (sys_session_start, sys_app_background, sys_sdk_health).
        occurred_at_client: { type: string, format: date-time }
        app_version: { type: string }
        build_number: { type: string }
        platform: { type: string }
        os_version: { type: string }
        device_class: { type: string }
        locale: { type: string }
        timezone_offset_minutes: { type: integer }
        properties: { $ref: "#/components/schemas/EventProperties" }

    BatchIngestRequest:
      type: object
      additionalProperties: false
      required: [schema_version, project_id, install_id, sdk, sent_at_client, events]
      properties:
        schema_version: { type: integer, const: 1 }
        project_id: { type: string }
        install_id: { type: string, format: uuid }
        sdk:
          type: object
          additionalProperties: false
          required: [name, version]
          properties:
            name: { type: string }
            version: { type: string }
        sent_at_client: { type: string, format: date-time }
        events:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: "#/components/schemas/Event" }

    RejectedEvent:
      type: object
      required: [event_id, code]
      properties:
        event_id: { type: string }
        code: { type: string }

    BatchIngestResponse:
      type: object
      required: [server_time, accepted, duplicates, rejected, client_policy]
      properties:
        server_time: { type: string, format: date-time }
        accepted: { type: array, items: { type: string } }
        duplicates: { type: array, items: { type: string } }
        rejected: { type: array, items: { $ref: "#/components/schemas/RejectedEvent" } }
        client_policy: { $ref: "#/components/schemas/ClientPolicy" }

    PolicyRequest:
      type: object
      additionalProperties: false
      required: [schema_version, project_id, install_id, sdk, app_version, build_number, platform]
      properties:
        schema_version: { type: integer, const: 1 }
        project_id: { type: string }
        install_id: { type: string, format: uuid }
        sdk:
          type: object
          additionalProperties: false
          required: [name, version]
          properties:
            name: { type: string }
            version: { type: string }
        app_version: { type: string }
        build_number: { type: string }
        platform: { type: string }

    PolicyResponse:
      type: object
      required: [server_time, client_policy]
      properties:
        server_time: { type: string, format: date-time }
        client_policy: { $ref: "#/components/schemas/ClientPolicy" }

    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    LoginResponse:
      type: object
      required: [role, project_ids, expires_at]
      properties:
        role: { type: string, enum: [admin, viewer] }
        project_ids: { type: array, items: { type: string } }
        expires_at: { type: string, format: date-time }

    SessionInfo:
      type: object
      required: [email, role, project_ids]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, viewer] }
        project_ids: { type: array, items: { type: string } }

    Overview:
      type: object
      description: See docs/metrics.md for every field's exact definition.
      required:
        - product_events, new_installations, daily_active_installations,
          weekly_active_installations, monthly_active_installations, sessions,
          avg_observed_session_duration_ms, ingestion_accepted, ingestion_duplicates, ingestion_rejected
      properties:
        product_events: { type: integer }
        new_installations: { type: integer }
        daily_active_installations: { type: integer }
        weekly_active_installations: { type: integer }
        monthly_active_installations: { type: integer }
        sessions: { type: integer }
        avg_observed_session_duration_ms: { type: number }
        ingestion_accepted: { type: integer }
        ingestion_duplicates: { type: integer }
        ingestion_rejected: { type: integer }

    EventExplorerResult:
      type: object
      required: [total_events, active_installations, trend]
      properties:
        total_events: { type: integer }
        active_installations: { type: integer }
        trend:
          type: array
          items:
            type: object
            required: [day, count]
            properties:
              day: { type: string, format: date }
              count: { type: integer }

    FunnelResult:
      type: object
      required: [steps, completion_window_seconds, truncated]
      properties:
        completion_window_seconds: { type: integer }
        truncated: { type: boolean, description: "true if the safety-valve event cap (500,000 rows) was hit — see internal/analytics/funnel.go." }
        steps:
          type: array
          items:
            type: object
            required: [name, count, conversion_from_first, conversion_from_previous]
            properties:
              name: { type: string }
              count: { type: integer }
              conversion_from_first: { type: number }
              conversion_from_previous: { type: number }

    RetentionResult:
      type: object
      required: [cohorts]
      properties:
        cohorts:
          type: array
          items:
            type: object
            required: [cohort_day, cohort_size, d1, d7, d30]
            properties:
              cohort_day: { type: string, format: date }
              cohort_size: { type: integer }
              d1: { type: integer }
              d7: { type: integer }
              d30: { type: integer }

    TimelineResult:
      type: object
      required: [install_id, registered_at, events, truncated]
      properties:
        install_id: { type: string, format: uuid }
        registered_at: { type: string, format: date-time }
        activated_at: { type: string, format: date-time }
        truncated: { type: boolean, description: "true if more than 500 events exist; only the most recent 500 are returned." }
        events:
          type: array
          items:
            type: object
            required: [event_id, name, event_kind, effective_at, time_quality, properties]
            properties:
              event_id: { type: string, format: uuid }
              name: { type: string }
              event_kind: { type: string, enum: [product, system] }
              effective_at: { type: string, format: date-time }
              time_quality: { type: string, enum: [client, batch_adjusted, untrusted] }
              properties: { type: object }

    CatalogResult:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            type: object
            required: [name, kind, description, owner, first_schema_version, properties, known]
            properties:
              name: { type: string }
              kind: { type: string, enum: [product, system] }
              description: { type: string }
              owner: { type: string }
              first_schema_version: { type: integer }
              properties: { type: array, items: { type: object } }
              known: { type: boolean, description: "false for auto-discovered events (empty description) — see section 7." }
              first_seen_at: { type: string, format: date-time }
              last_seen_at: { type: string, format: date-time }

    SystemHealth:
      type: object
      required:
        - version, db_latency_ms, writer_pool, reader_pool, disk_state,
          ingestion_accepted_last_hour, ingestion_rejected_last_hour,
          enabled_policy_rules, last_maintenance_runs
      properties:
        version: { type: string }
        db_latency_ms: { type: number }
        disk_state: { type: string, enum: [normal, warning, high, critical, rejecting] }
        ingestion_accepted_last_hour: { type: integer }
        ingestion_rejected_last_hour: { type: integer }
        enabled_policy_rules: { type: integer }
        writer_pool: { $ref: "#/components/schemas/PoolStats" }
        reader_pool: { $ref: "#/components/schemas/PoolStats" }
        last_maintenance_runs:
          type: array
          items:
            type: object
            required: [kind, started_at, rows_affected]
            properties:
              kind: { type: string }
              started_at: { type: string, format: date-time }
              finished_at: { type: string, format: date-time }
              rows_affected: { type: integer }
              error: { type: string }

    PoolStats:
      type: object
      required: [acquired_conns, idle_conns, total_conns, max_conns]
      properties:
        acquired_conns: { type: integer }
        idle_conns: { type: integer }
        total_conns: { type: integer }
        max_conns: { type: integer }

    PolicyRule:
      type: object
      required: [id, project_id, mode, next_check_seconds, discard_pending, reason, enabled, created_at, updated_at]
      properties:
        id: { type: integer }
        project_id: { type: string }
        environment: { type: string }
        app_version: { type: string }
        build_number: { type: string }
        sdk_version: { type: string }
        mode: { type: string, enum: [active, pause_upload, disable_collection] }
        next_check_seconds: { type: integer }
        discard_pending: { type: boolean }
        reason: { type: string }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    PolicyCreateRequest:
      type: object
      additionalProperties: false
      required: [project_id, mode, next_check_seconds, reason]
      properties:
        project_id: { type: string }
        environment: { type: string }
        app_version: { type: string }
        build_number: { type: string }
        sdk_version: { type: string }
        mode: { type: string, enum: [active, pause_upload, disable_collection] }
        next_check_seconds: { type: integer }
        discard_pending: { type: boolean }
        reason: { type: string, description: "Required — every kill-switch action is audited with its reason (section 5.5)." }

  parameters:
    Project:
      name: project
      in: query
      required: true
      schema: { type: string }
    From:
      name: from
      in: query
      description: RFC 3339. Defaults to 7 days before `to`.
      schema: { type: string, format: date-time }
    To:
      name: to
      in: query
      description: RFC 3339. Defaults to now. Window (to - from) cannot exceed 90 days.
      schema: { type: string, format: date-time }
    Timezone:
      name: timezone
      in: query
      description: IANA timezone name. Defaults to UTC.
      schema: { type: string }

  responses:
    Error:
      description: Stable machine-readable error. See docs/errors.md for the code list and retry classification.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }

paths:
  /v1/installs/register:
    post:
      summary: Idempotent anonymous installation registration
      description: See server_implementation_plan.md section 5.2 for the atomic-behavior rules (unknown ID, matching retry, credential conflict).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: Registered, or a lost-response retry of an identical prior registration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "409":
          description: install_conflict — install_id already registered with a different credential.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }

  /v1/events/batch:
    post:
      summary: Authenticated gzip batch event ingestion
      security:
        - installationBearer: []
      description: See server_implementation_plan.md section 5.4 for limits and the accept/duplicate/reject transaction rules.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchIngestRequest" }
      responses:
        "200":
          description: Processed. Individual events may still be listed as rejected — check the rejected array, not just the status code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchIngestResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503":
          description: server_storage_pressure or maintenance. Includes Retry-After.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /v1/client/policy:
    post:
      summary: Lightweight authenticated policy probe (no events)
      security:
        - installationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "400": { $ref: "#/components/responses/Error" }

  # --- Dashboard API (section 10.1) ---------------------------------------
  # Auth + Overview/Event Explorer are fully specified (Phase S2).
  # funnel/retention/timeline/catalog/system remain path stubs until Phase
  # S3's read-only query surface exists for them.

  /api/v1/auth/login:
    post:
      summary: Admin/viewer login (section 10.3)
      security: []
      description: >
        No public account creation — an admin_users row only ever comes
        from the create-admin CLI. Throttled per email and per source IP;
        the error is identical whether the account doesn't exist, is
        disabled, or the password is wrong.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Sets the session (HttpOnly) and csrf_token (JS-readable) cookies.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }

  /api/v1/auth/logout:
    post:
      summary: Invalidate the current dashboard session
      security: [{ dashboardSession: [] }]
      description: Requires the X-CSRF-Token header to match the csrf_token cookie (double-submit, section 10.3). Idempotent.
      responses:
        "204": { description: Session revoked (or was already invalid). }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/auth/session:
    get:
      summary: Current dashboard session identity and role
      security: [{ dashboardSession: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionInfo" }
        "401": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/overview:
    get:
      summary: Overview screen metrics (section 10.2 #1)
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Overview" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/events:
    get:
      summary: Event explorer trends (section 10.2 #2)
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Timezone"
        - name: name
          in: query
          description: Must be a name in the project's event catalog (or a reserved system event).
          schema: { type: string }
        - name: app_version
          in: query
          schema: { type: string }
        - name: build_number
          in: query
          schema: { type: string }
        - name: platform
          in: query
          schema: { type: string }
        - name: property_key
          in: query
          description: Requires `name`; must be declared in that event's catalog entry. Requires property_value.
          schema: { type: string }
        - name: property_value
          in: query
          description: Requires property_key.
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventExplorerResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/funnel:
    get:
      summary: Funnel conversion query (section 10.2 #3)
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: steps
          in: query
          required: true
          description: 2 to 5 comma-separated product event names, each must be in the project's catalog.
          schema: { type: string }
        - name: window_seconds
          in: query
          description: Completion window measured from the first step's occurrence. Defaults to 3600, max 86400.
          schema: { type: integer }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FunnelResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/retention:
    get:
      summary: D1/D7/D30 installation retention cohorts (section 10.2 #4)
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RetentionResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/installations/{id}:
    get:
      summary: Single installation event timeline, admin-only (section 10.2 #5)
      security: [{ dashboardSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Project"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineResult" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: Not admin role, or not scoped to this project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { $ref: "#/components/responses/Error" }

  /api/v1/analytics/catalog:
    get:
      summary: Event catalog definitions and known/unknown status (section 10.2 #6)
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogResult" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/system:
    get:
      summary: System health (section 10.2 #7)
      description: DB/pool/disk state are cluster-wide; ingestion rate and policy-rule counts are scoped to the session's accessible projects.
      security: [{ dashboardSession: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemHealth" }
        "401": { $ref: "#/components/responses/Error" }

  /api/v1/policy:
    get:
      summary: List kill-switch policy rules for a project
      security: [{ dashboardSession: [] }]
      parameters:
        - $ref: "#/components/parameters/Project"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules: { type: array, items: { $ref: "#/components/schemas/PolicyRule" } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      summary: Create a kill-switch policy rule (admin-only)
      security: [{ dashboardSession: [] }]
      description: Requires the X-CSRF-Token header (double-submit, section 10.3). Every rule requires a reason — this is an audited action.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyCreateRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyRule" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

  /api/v1/policy/{id}:
    delete:
      summary: Remove a kill-switch policy rule (admin-only)
      security: [{ dashboardSession: [] }]
      description: Requires the X-CSRF-Token header (double-submit, section 10.3).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/Project"
      responses:
        "204": { description: Rule removed. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }